+++ src/LightOps.Commerce.Services.Product/Api/Queries/FetchProductsByCategoryIdsQuery.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using LightOps.CQRS.Api.Queries;
+
+namespace LightOps.Commerce.Services.Product.Api.Queries
+{
+    public class FetchProductsByCategoryIdsQuery : IQuery
+    {
+        public IList<string> CategoryIds { get; set; }
+    }
+}